FIX KB-26: Moved page layout to `/pages/`
